Added Dockerfile, docker-cmpose for deploy, controller refactor

+++ internal/comment/Controller.go
@@ -2,8 +2,8 @@ package comment
 
 import (
 	"encoding/json"
-	"github.com/roobiuli/MongoCRUD/internal/pkg/storage"
 	"github.com/google/uuid"
+	"github.com/roobiuli/MongoCRUD/internal/pkg/storage"
 	"net/http"
 	"time"
 )
@@ -17,6 +17,18 @@ type Controller struct {
 
 
 func(c Controller) Create(w http.ResponseWriter, r *http.Request) {
+
+	if r.Method == http.MethodPost {
+		w.WriteHeader(http.StatusMethodNotAllowed)
+		w.Write([]byte("POST Method required"))
+		return
+	}
+
+	if r.Method == http.MethodGet {
+		w.WriteHeader(http.StatusMethodNotAllowed)
+		w.Write([]byte("GET Method required"))
+		return
+	}
 
 	var req Create
 
@@ -50,6 +62,12 @@ func(c Controller) Create(w http.ResponseWriter, r *http.Request) {
 
 
 func(c Controller) Find(w http.ResponseWriter, r *http.Request) {
+	if r.Method == http.MethodGet {
+		w.WriteHeader(http.StatusMethodNotAllowed)
+		w.Write([]byte("GET Method required"))
+		return
+	}
+
 	id := r.Form.Get("uuid")
 
 	 com, err := c.Storage.Find(r.Context(), id)
@@ -82,6 +100,12 @@ func(c Controller) Find(w http.ResponseWriter, r *http.Request) {
 }
 
 func (c Controller) Delete(w http.ResponseWriter, r *http.Request)  {
+	if r.Method == http.MethodDelete{
+		w.WriteHeader(http.StatusMethodNotAllowed)
+		w.Write([]byte("Delete Method required"))
+		return
+	}
+
 	id :=  r.Form.Get("uuid")
 
 	if err := c.Storage.Delete(r.Context(), id); err != nil {
@@ -96,6 +120,12 @@ func (c Controller) Delete(w http.ResponseWriter, r *http.Request)  {
 }
 
 func (c Controller) Update(w http.ResponseWriter, r *http.Request) {
+
+	if r.Method == http.MethodPatch {
+		w.WriteHeader(http.StatusMethodNotAllowed)
+		w.Write([]byte("PATCH Method required"))
+		return
+	}
 	var req Create
 	id := r.Form.Get("uuid")
 	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
